refactor(reservation-app): 예약 흐름의 Seat 포트 참조 변경
- ReservationCommandService가 Room 패키지로 이동한 SeatStore를 참조하도록 변경
- 예약, 대기열, 동시성 통합 테스트의 SeatCommandService와 CreateSeatCommand import를 갱신
- 테스트 DB 정리 코드가 이동된 SeatRepository를 참조하도록 변경

- 테스트:
  - 좌석 생성 결과 타입 변경에 맞춰 통합 테스트 검증 갱신
