+++ reservation/reservation-application/src/test/java/com/seatliberator/seatliberator/reservation/integration/ReservationConcurrencyTest.java
@@ -2,8 +2,8 @@ package com.seatliberator.seatliberator.reservation.integration;
 
 import com.seatliberator.seatliberator.reservation.book.application.port.in.command.CreateReservationCommand;
 import com.seatliberator.seatliberator.reservation.book.application.service.ReservationCommandService;
-import com.seatliberator.seatliberator.reservation.seat.application.port.in.command.CreateSeatCommand;
-import com.seatliberator.seatliberator.reservation.seat.application.service.SeatCommandService;
+import com.seatliberator.seatliberator.reservation.room.application.port.in.command.CreateSeatCommand;
+import com.seatliberator.seatliberator.reservation.room.application.service.SeatCommandService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Tag;
