+++ reservation/reservation-application/src/test/java/com/seatliberator/seatliberator/reservation/integration/ReservationDatabaseCleanupSupport.java
@@ -1,7 +1,7 @@
 package com.seatliberator.seatliberator.reservation.integration;
 
 import com.seatliberator.seatliberator.reservation.book.infrastructure.persistence.jpa.repository.ReservationRepository;
-import com.seatliberator.seatliberator.reservation.seat.infrastructure.persistence.jpa.repository.SeatRepository;
+import com.seatliberator.seatliberator.reservation.room.infrastructure.persistence.jpa.repository.SeatRepository;
 import com.seatliberator.seatliberator.reservation.waitlist.infrastructure.persistence.jpa.repository.WaitlistRepository;
 import org.junit.jupiter.api.AfterEach;
 import org.springframework.beans.factory.annotation.Autowired;
