+++ reservation/reservation-application/src/test/java/com/seatliberator/seatliberator/reservation/integration/WaitlistRequestTest.java
@@ -4,8 +4,8 @@ import com.seatliberator.seatliberator.reservation.WaitlistCreateCommandBuilder;
 import com.seatliberator.seatliberator.reservation.book.application.port.in.CreateReservationUseCase;
 import com.seatliberator.seatliberator.reservation.book.application.port.in.command.CreateReservationCommand;
 import com.seatliberator.seatliberator.reservation.domain.WaitlistStatus;
-import com.seatliberator.seatliberator.reservation.seat.application.port.in.command.CreateSeatCommand;
-import com.seatliberator.seatliberator.reservation.seat.application.service.SeatCommandService;
+import com.seatliberator.seatliberator.reservation.room.application.port.in.command.CreateSeatCommand;
+import com.seatliberator.seatliberator.reservation.room.application.service.SeatCommandService;
 import com.seatliberator.seatliberator.reservation.shared.application.exception.ReservationApplicationException;
 import com.seatliberator.seatliberator.reservation.waitlist.application.port.in.CancelWaitlistUseCase;
 import com.seatliberator.seatliberator.reservation.waitlist.application.port.in.CreateWaitlistUseCase;
